♻️ Split details page into component (as gf)
